requirements.txt
-e .
aws-cdk.aws_codecommit
aws-cdk.aws_codepipeline
aws-cdk.aws_codepipeline_actions
aws-cdk.pipelines
aws-cdk.aws_codestarconnections
aws-cdk.aws_s3
aws-cdk.aws_config
aws-cdk.aws_guardduty
aws-cdk.aws_macie
aws-cdk.aws_accessanalyzer
boto3